refactor: move interfaces 'comands' and 'events' into discord folder
