+++ src/application/commands/development/test-welcome.ts
@@ -1,4 +1,4 @@
-import { Command } from "@interfaces/commands/Command";
+import { Command } from "@interfaces/discord/Command";
 import { ApplicationCommandType } from "discord.js";
 
 export default new Command({
