+++ src/domain/interfaces/discord/Command.ts
