+++ src/domain/interfaces/discord/event.ts
